refactor: removing ui folder.
