+++ lib/src/app_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import 'presenter/shared/ui/theme/theme_config.dart';
-import 'presenter/shared/ui/routes/app_route_manager.dart';
+import 'presenter/shared/extentions/theme/theme_config.dart';
+import 'presenter/shared/routes/app_route_manager.dart';
 
 class RachaRachaApp extends StatelessWidget {
   const RachaRachaApp({Key? key}) : super(key: key);
