+++ lib/src/presenter/shared/widgets/custom_will_pop_scope_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../ui/constants/space_constants.dart';
+import '../constants/space_constants.dart';
 import '../controllers/check_controller.dart';
 import '../utils/custom_utils.dart';
